requirements.txt
torch
torchvision
faiss-cpu
opencv-python
pandas
numpy
streamlit
pillow
pytesseract